refactor: Decouple GamePage Layout from PacmanGame
Decoupling the game logic handling component from the
purely layout component for displaying the game
