+++ client/src/components/PacmanGame/PacmanGame.jsx
@@ -1,10 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import Paper from '@material-ui/core/Paper';
-import Grid from '@material-ui/core/Grid';
-import Button from '@material-ui/core/Button';
 import pathfinding from 'pathfinding';
+import GamePage from '../Layout/GamePage';
 
 import {
   initSquareGridState,
@@ -18,16 +15,6 @@ import {
   entityApplier,
 } from './gameCore';
 import PacmanBoard from './PacmanBoard';
-import ScoreCard from '../UI/ScoreCard';
-
-const styles = theme => ({
-  root: {
-    ...theme.mixins.gutters(),
-    paddingTop: theme.spacing.unit * 2,
-    paddingBottom: theme.spacing.unit * 2,
-    width: '96%',
-  },
-});
 
 class PacmanGame extends Component {
   state = {
@@ -282,36 +269,19 @@ class PacmanGame extends Component {
     const {
       gridState, pacman, score, status,
     } = this.state;
-    const { classes } = this.props;
     return (
-      <React.Fragment>
-        <Paper className={classes.root} elevation={1}>
-          <Grid container spacing={16}>
-            <Grid item xs={3} />
-            <Grid item xs={6}>
-              <Button variant="outlined" size="medium" color="primary" onClick={this.startGame}>
-                Start
-              </Button>
-              <div>
-                Score:
-                {score}
-              </div>
-              <div>
-                Status:
-                {status === 2 ? 'GAME OVER' : status}
-              </div>
-              <PacmanBoard
-                gridSize={gridSize}
-                gridState={gridState}
-                pacman={pacman}
-              />
-            </Grid>
-            <Grid item xs={3} sm container>
-              <ScoreCard />
-            </Grid>
-          </Grid>
-        </Paper>
-      </React.Fragment>
+      <GamePage
+        startGame={this.startGame}
+        score={score}
+        status={status}
+        render={() => (
+          <PacmanBoard
+            gridSize={gridSize}
+            gridState={gridState}
+            pacman={pacman}
+          />
+        )}
+      />
     );
   }
 }
@@ -319,7 +289,6 @@ class PacmanGame extends Component {
 PacmanGame.propTypes = {
   width: PropTypes.number.isRequired,
   numberofCells: PropTypes.number.isRequired,
-  classes: PropTypes.shape().isRequired,
 };
 
-export default withStyles(styles)(PacmanGame);
+export default PacmanGame;
